Separete `<CircleProgress />`
