+++ app/javascript/mastodon/components/dropdown_menu.jsx
@@ -8,7 +8,7 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import { supportsPassiveEvents } from 'detect-passive-events';
 import Overlay from 'react-overlays/Overlay';
 
-import { CircularProgress } from 'mastodon/components/loading_indicator';
+import { CircularProgress } from 'mastodon/components/circle_progress';
 
 import { IconButton } from './icon_button';
